+++ lib/presentation/screens/home_screen.dart
@@ -5,7 +5,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 // 🌎 Project imports:
-import 'package:kreator_frame/l10n/app_localizations.dart';
 import 'package:kreator_frame/presentation/providers/providers.dart';
 import 'package:kreator_frame/presentation/widgets/widgets.dart';
 
@@ -25,7 +24,6 @@ class HomeScreen extends ConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final tabsBar = ref.watch(tabsBarAppProvider);
-    final textStyles = Theme.of(context).textTheme;
 
     // Auto-execute immediate update when available
     ref.listen(inAppUpdateProvider, (previous, next) async {
@@ -34,53 +32,31 @@ class HomeScreen extends ConsumerWidget {
       }
     });
 
-    return tabsBar.when(
-      data: (data) {
-
-        return Scaffold(
-          body: DefaultTabController(
-            length: data.length,
-            child: Builder(builder: (context) {
-              return NestedScrollView(
-                headerSliverBuilder: (context, innerBoxIsScrolled) {
-                  return [
-                    // App Bar
-                    const CustomSliverAppBar(),        
-                  ];
-                },
-                body: TabBarView(
-                  children: data.map((tabEntity) => tabEntity.tabBarView).toList(),
-                ),
-              );
-            }),
-          ),
-        );
-
-      }, 
-      error: (error, stackTrace) {
-
-        return Scaffold(
-          body: Center(
-            child: Text(
-              AppLocalizations.of(context)!.errorMessage,
-              style: textStyles.titleLarge,
-            ),
-          ),
-        );
-        
-      }, 
-      loading: () {
-        
-        return const Scaffold(
-          body: Center(
-            child: CircularProgressIndicator(
-              strokeCap: StrokeCap.round
-            )
-          ),
-        );
-
-      },
+    return Scaffold(
+      body: tabsBar.when(
+        data: (data) => DefaultTabController(
+          length: data.length,
+          child: Builder(builder: (context) {
+            return NestedScrollView(
+              headerSliverBuilder: (context, innerBoxIsScrolled) {
+                return [
+                  // App Bar
+                  const CustomSliverAppBar(),        
+                ];
+              },
+              body: TabBarView(
+                children: data.map((tabEntity) => tabEntity.tabBarView).toList(),
+              ),
+            );
+          }),
+        ), 
+        error: (_, _) => ErrorView(
+          onRetry: () => ref.invalidate(tabsBarAppProvider),
+        ), 
+        loading: () => const Center(
+          child: CircularProgressIndicator(strokeCap: StrokeCap.round)
+        ),
+      ),
     );
-
   }
 }